Assignment 1 core.
In this assingment, allocations are maintained using a free list. This free list
is an extension of the python list class, with additional functionality for
pushing and popping free list nodes while maintaining a first-fit ordering. The
free list consists solely of FreeListNodes, which are simple data objects
containing only a heap_index and size.

All statistics are generated by the FreeListStatsGenerator class, which deals
with lists of FreeListNodes.