Major restructuring. Memory Manager introduced.
'first-fit' behavior has been de-coupled from the freeList, leaving freeList as
a simple python list. 'first-fit' behavior is now contained in the
FirstFitAllocator class.

A Memory Manager (intermediate layer) between the application and the heap has
been introduced. Rather than have the application communicate directly with the
heap, the MemoryManager class is an abstraction which takes care of allocation
and delegation. The heap (actual data) has been reduced to a simple byte-array,
and will be a field of the memory manager if it turns out that we actually need
the heap data.

Allocators and Garbage Collectors are to be designed as plugins for the
MemoryManager and are specified by dependency injection when instantiating the
MemoryManager.