Heap can now be configured to grow by a factor. TODO
Simulator can be configured with 2 additional parameters:
	- high_water_percent
	- growth_factor

Immediately after a collection occurs, if the percentage of heap occupied is
greater than <high_water_percent>, the heap will be expanded by a factor of
<growth_factor>.

For example, consider a heap of 100 bytes with:
	- high_water_percent = 50
	- growth_factor = 2.0

Immediately after a collection there are still 61 bytes occupied by live objects
(i.e. 61% occupied). In this scenario, the heap will be expanded by a factor of
2, and will now be 200 bytes in size.

TODO: this commit also contains collector debugging output. I need to do
something with that. It doesn't belog here.