Added pluggable write-barriers.
The simulator can now be configured with write barriers. These write barriers
take the form of a class with a single process() method which accepts the:
	- source object id
	- old destination object id (where the pointer currently points)
	- new destination object id (where the pointer is about to point)

A list of write barriers can be injected into the memory manager object from the
main driver script. All of the write barriers process() methods will be invoked
on each pointer write operation, and additionally on each rootset addition and
removal.

This structure will be useful for performing Reference Counting GC, as well as
for other statistics-gathering purposes.