Added Recycler collector configuration.
The recycler is a reference counting GC policy which resolves cycles by only
performing localized, partial traversals. This efficient cycle resolution is
based on the insight that only objects whose most recent reference count update
was a decrement that resulted in a non-zero reference count. Objects that that
satisfy this property are known as 'candidates'. When the heap becomes full, a
trial deletion is performed over the transitive closure of these 'candidates'.

The recycler configuration can be invoked using the following command line
options (the 2nd 'recycler' specifies the collector):

	--write-barrier recycler recycler

Note: these policies are highly coupled. While the recycler write-barrier can be
paired with a traversal collector, the reverse is not true. Pairing the recycler
collector with any other write-barrier (or lack-there-of) will result in
undefined behavior.

Note: this implementation of recycler differs from the implementation in the GC
Handbook. In the book, once an object is identified as a candidate, its
collection is postponed until the next collection, even if subsequent reference
count decrements results in a RC of 0. In this implementation, an object in this
situation would be collected immediately.