Optimization. _merge_adjacent_nodes() no longer traverses freelist.
The _merge_adjacent_nodes() sub-routine was unnecessarily traversing the entire
free list on each invocation. This sub-routine is called every time a new node
is added to the free list.

Since we know the index of the newly added node, this traversal is unnecessary.
Instead, we can simply check to see if the newly added node is adjacent to its
two surrounding neighbours.