Added a 'next fit' allocation policy.
This allocation policy caches the location of the last successful allocation,
and starts its search for the next allocation beginning from this point.

The motivation for this policy is that over time, very small and fragmented
blocks of memory settle at the beginning of the heap. This policy avoids
repeatedly traversing these 'useless' blocks of memory, and should therefore
result in more efficient allocation.